refactor(casper): consolidate test node fixtures
- centralize node, transport, genesis, and Rholang test resources
- retain compatibility re-exports and standalone behavior coverage
- record the ratified B1 characterization baseline
