+++ casper/src/rust/test_utils/util/comm/transport_layer_test_impl.rs
@@ -1,6 +1,4 @@
 // Rust port of casper/src/test/scala/coop/rchain/casper/util/comm/TransportLayerTestImpl.scala
-// Moved from casper/tests/util/comm/transport_layer_test_impl.rs to casper/src/rust/test_utils/util/comm/transport_layer_test_impl.rs
-// No import changes needed - this file doesn't use casper:: or crate::util::
 
 use std::collections::{HashMap, VecDeque};
 use std::future::Future;
@@ -38,12 +36,13 @@ pub mod test_network {
             }
         }
 
-        /// Add a peer to the network with an empty message queue
+        /// Add a peer to the network with an empty message queue (only if not already present)
         pub fn add_peer(&self, peer: &PeerNode) -> Result<(), CommError> {
             let mut state = self.state.lock().map_err(|_| {
                 CommError::InternalCommunicationError("Failed to acquire state lock".to_string())
             })?;
-            state.insert(peer.clone(), VecDeque::new());
+            // Only add if peer doesn't exist - don't replace existing queue!
+            state.entry(peer.clone()).or_insert_with(VecDeque::new);
             Ok(())
         }
 
@@ -134,13 +133,6 @@ impl TransportLayerTestImpl {
     /// Create a new test transport layer with the given test network
     pub fn new(test_network: test_network::TestNetwork) -> Self { Self { test_network } }
 
-    /// Create a new test transport layer with an empty test network
-    pub fn empty() -> Self {
-        Self {
-            test_network: test_network::TestNetwork::empty(),
-        }
-    }
-
     /// Get access to the underlying test network for test setup
     pub fn test_network(&self) -> &test_network::TestNetwork { &self.test_network }
 }
@@ -164,7 +156,7 @@ impl TransportLayer for TransportLayerTestImpl {
     }
 
     async fn stream(&self, peer: &PeerNode, blob: &Blob) -> Result<(), CommError> {
-        self.stream_mult(&[peer.clone()], blob).await
+        self.stream_mult(std::slice::from_ref(peer), blob).await
     }
 
     async fn stream_mult(&self, peers: &[PeerNode], blob: &Blob) -> Result<(), CommError> {
@@ -175,9 +167,13 @@ impl TransportLayer for TransportLayerTestImpl {
         self.broadcast(peers, &protocol_msg).await
     }
 
-    async fn disconnect(&self, _peer: &PeerNode) -> Result<(), CommError> { Ok(()) }
+    async fn disconnect(&self, _peer: &PeerNode) -> Result<(), CommError> {
+        // Test implementation - do nothing
+        Ok(())
+    }
 
     async fn get_channeled_peers(&self) -> Result<std::collections::HashSet<PeerNode>, CommError> {
+        // Test implementation - return empty set
         Ok(std::collections::HashSet::new())
     }
 }
@@ -205,23 +201,18 @@ impl TransportLayerServer for TransportLayerServerTestImpl {
         dispatch: DispatchFn,
         _handle_streamed: HandleStreamedFn,
     ) -> Result<Cancelable, CommError> {
-        let identity = self.identity.clone();
-        let test_network = self.test_network.clone();
+        // In tests, process all messages synchronously before returning
+        let dispatch_fn = move |protocol: Protocol| {
+            let dispatch = dispatch.clone();
+            async move { dispatch(protocol).await }
+        };
 
-        // Create a long-running task that processes messages from the queue
-        let handle = tokio::spawn(async move {
-            // Create a dispatch function that matches the expected signature
-            let dispatch_fn = move |protocol: Protocol| {
-                let dispatch = dispatch.clone();
-                async move { dispatch(protocol).await }
-            };
+        // Process all messages in the queue for this peer synchronously
+        self.test_network
+            .handle_queue(dispatch_fn, &self.identity)
+            .await?;
 
-            // Process all messages in the queue for this peer
-            if let Err(e) = test_network.handle_queue(dispatch_fn, &identity).await {
-                tracing::error!(peer = %identity, error = %e, "test transport queue processing failed");
-            }
-        });
-
-        Ok(handle)
+        // Return a no-op cancelable since we've already processed everything
+        Ok(tokio::spawn(async {}))
     }
 }
