+++ casper/tests/util/comm/mod.rs
@@ -1,2 +1,1 @@
 pub mod fair_round_robin_dispatcher_test;
-pub mod transport_layer_test_impl;
